Split up io test cases (results vs input data)
